+++ rte/src/brux/main.cpp
@@ -255,22 +255,22 @@ void xyEnd(){
 	//Cleanup all resources
 	xyPrint(0, "Cleaning up all resources...");
 	xyPrint(0, "Cleaning textures...");
-	for(int i = 0; i < vcTextures.size(); i++) {
+	for(int i = 0; i < static_cast<int>(vcTextures.size()); i++) {
 		xyDeleteImage(i);
 	}
 
 	xyPrint(0, "Cleaning sprites...");
-	for(int i = 0; i < vcSprites.size(); i++) {
+	for(int i = 0; i < static_cast<int>(vcSprites.size()); i++) {
 		delete vcSprites[i];
 	}
 
 	xyPrint(0, "Cleaning sounds...");
-	for(int i = 0; i < vcSounds.size(); i++) {
+	for(int i = 0; i < static_cast<int>(vcSounds.size()); i++) {
 		xyDeleteSound(i);
 	}
 
 	xyPrint(0, "Cleaning music...");
-	for(int i = 0; i < vcMusic.size(); i++) {
+	for(int i = 0; i < static_cast<int>(vcMusic.size()); i++) {
 		xyDeleteMusic(i);
 	}
 	xyPrint(0, "Finished cleanup.");
